Update minitests to work with multiple VS versions
