+++ files/default/test/install_update_test.rb
@@ -5,12 +5,13 @@ class InstallUpdateSpec < MiniTest::Chef::Spec
 
   describe_recipe 'visualstudio::install_update' do
     describe 'packages' do
+      include ChefHelper
 
-      it 'installs VS 2012 SP4' do
-        node['visualstudio']['update']['package_name'].must_be_installed
+      it 'installs Visual Studio update' do
+        each_version_edition(node) do |version, _|
+          node['visualstudio'][version]['update']['package_name'].must_be_installed
+        end
       end
-
     end
   end
-
 end
